requirements.txt
streamlit
Pillow==9.5.0
opencv-python==4.8.0.76
ultralytics
split-folders
